Added sp call for UserInfo Get Added sp call for Group Get Added sp call for Group Create

+++ src/app/Chaos.Portal/Data/Dto/IUserInfo.cs
@@ -14,7 +14,7 @@
         Guid? SessionGuid { get; set; }
 
         [Serialize]
-        long? SystemPermissions { get; set; }
+        uint? SystemPermissions { get; set; }
         SystemPermissons SystemPermissonsEnum { get; set; }
 
         [Serialize]
